refactor and move search sorted logic
